+++ packages/docs/site/src/routes/components/_examples/index.js
@@ -2,4 +2,4 @@ export {default as barchart} from './barchart.js';
 export {default as choropleth} from './choropleth.js';
 export {default as histogram} from './histogram.js';
 export {default as legend} from './legend.js';
-export {default as ui} from './ui.js';
+export {default as ui} from './ui/index.js';