+++ packages/docs/site/src/routes/components/_examples/ui/index.js
@@ -1,0 +1,21 @@
+import * as AlphabetPicker from './AlphabetPicker.js';
+import * as CopyToClipboard from './CopyToClipboard.js';
+import * as HighlightedText from './HighlightedText.js';
+import * as Icon from './Icon.js';
+import * as Link from './Link.js';
+import * as LinkButton from './LinkButton.js';
+import * as LoadingView from './LoadingView.js';
+import * as Switch from './Switch.js';
+
+import {formatExamples} from '../utils.js';
+
+export default formatExamples([
+	AlphabetPicker,
+	CopyToClipboard,
+	HighlightedText,
+	Icon,
+	Link,
+	LinkButton,
+	LoadingView,
+	Switch,
+]);